fix: replace nint with IntPtr for C# 7.3 compatibility and fix enum conversions

+++ src/Andastra/Runtime/Graphics/Stride/Audio/StrideSoundPlayer.cs
@@ -579,7 +579,7 @@ namespace Andastra.Runtime.Stride.Audio
                     try
                     {
                         IntPtr bufferPtr = silenceHandle1.AddrOfPinnedObject();
-                        FillBuffer(bufferPtr, bufferSizeBytes, (Stride.Audio.AudioLayer.BufferType)AudioLayer.BufferType.EndOfStream);
+                        FillBuffer(bufferPtr, bufferSizeBytes, (Stride.Audio.AudioLayer.BufferType)(int)AudioLayer.BufferType.EndOfStream);
                     }
                     finally
                     {
@@ -736,7 +736,7 @@ namespace Andastra.Runtime.Stride.Audio
                     try
                     {
                         IntPtr bufferPtr = silenceHandle3.AddrOfPinnedObject();
-                        FillBuffer(bufferPtr, bufferSizeBytes, (Stride.Audio.AudioLayer.BufferType)AudioLayer.BufferType.EndOfStream);
+                        FillBuffer(bufferPtr, bufferSizeBytes, (Stride.Audio.AudioLayer.BufferType)(int)AudioLayer.BufferType.EndOfStream);
                     }
                     finally
                     {
@@ -771,7 +771,7 @@ namespace Andastra.Runtime.Stride.Audio
             try
             {
                 IntPtr bufferPtr = bufferHandle.AddrOfPinnedObject();
-                FillBuffer(bufferPtr, bytesToCopy, (Stride.Audio.AudioLayer.BufferType)bufferType);
+                FillBuffer(bufferPtr, bytesToCopy, (Stride.Audio.AudioLayer.BufferType)(int)bufferType);
             }
             finally
             {
